Synchroniser + OptoStim First Trial
Includes optoStim + synchroniser sketch written by Mike Payne. This sketch sends one pseudo-random pulse to pin A11 to be used with the KitchenPulse synchroniser system. This sketch also sends an optical stimulation pulse to pin A14. This pulse will only turn on when one of three buttons is pressed. The signal from the first button is fed into pin 3 and will start an optical pulse that is 0.5 Hz with a 10 msec pulse. The second button is fed into pin 5 and is 0.5 Hz. with a 50 msec pulse. The third button is 0.5 Hz. with a 100 msec pulse.