Combines Rewritten Kitchen Sync with OptoStim
This sketch combines the rewritten kitchen sync code with code used to deliver optical sitmulation. The kitchen sync code runs continuously as long as the Arduino is on and is used to synchronize two functions (in our case, the frames of the camera with the acquired Intan ephys signal). The optotagging feature will only turn on when one of three buttons is pushed. Each button is programmed to deliver a different type of optical pulse.